Streamlit/progress/requirements.txt
streamlit
pandas
numpy
plotly
mysql-connector-python
oracledb